Placed date input into a separate component to avoid entire header rerender
